requirements.txt
arrow==1.2.3
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.0
binaryornot==0.4.4
black==22.10.0
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
cssbeautifier==1.14.7
dill==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-admin-tailwind==0.0.16
django-admin-volt==1.0.4
django-appconf==1.0.5
django-browser-reload==1.6.0
django-components==0.22
django-compressor==4.1
django-debug-toolbar==3.7.0
django-tailwind==3.4.0
djlint==1.19.7
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==5.0.0
isort==5.10.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsbeautifier==1.14.7
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
pep8==1.7.1
Pillow==9.3.0
platformdirs==2.5.4
psycopg2==2.9.5
pycodestyle==2.9.1
pylint==2.15.6
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
rcssmin==1.1.0
regex==2022.10.31
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0
wrapt==1.14.1
xlwt==1.3.0
zipp==3.10.0
